requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
django-cors-headers==4.5.0
django-ninja==1.3.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozenlist==1.5.0
google-ai-generativelanguage==0.6.10
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
grpcio==1.67.0
grpcio-status==1.67.0
httplib2==0.22.0
idna==3.10
multidict==6.1.0
packaging==24.1
pillow==11.0.0
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pyparsing==3.2.0
PyPDF2==3.0.1
pytesseract==0.3.13
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
sqlparse==0.5.1
tqdm==4.66.5
twilio==9.3.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
yarl==1.16.0
